Sessions now have a proper queue: every recording gets added to the queue of everyone.
There is an admin option (#), but only if you're on a caller id
whitelist.  From the admin, you can queue audio files audio/0 until
audio/9 by pressing the right digit.

The files added via the admin menu get added to the head of the queue,
so they'll be played very soon on all sessions.